services/load-testing/requirements.txt
locustio~=0.11
